+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 01. :gift: Added interface `dev.orne.config.ConfigCryptoProvider`
     01. :gift: Added exception `dev.orne.config.ConfigCryptoProviderException`
 01. :gift: Added class `dev.orne.config.EncryptedConfig`
+01. :gift: Added class `dev.orne.config.MutableEncryptedConfig`
 
 # :package: 0.1.0